pyproject.toml
[project]
name = "hermes"
version = "1.0.0"
description = "Weather→Markets Trading System (Polymarket-first, Kalshi-ready)"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"

dependencies = [
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "pytz>=2024.1",
    "python-dateutil>=2.8.2",
    "pandas>=2.1.0",
    "numpy>=1.24.0",
    "scipy>=1.11.0",
    "tenacity>=8.2.0",
    "websockets>=12.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["core*", "agents*", "venues*"]
exclude = ["tests*", "data*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

